rollout depthformer : hidden suit le dtype du modèle (mismatch fp16 vu sur T4)
La machine à états interleaved s'active en runtime réel : sample() bascule
bien en AUDIO après le bloc texte — le rollout plantait sur depth_linear
(hidden cast en float32 vs poids fp16).

+++ src/vllm_omni_lfm2_audio/lfm2_audio_ar.py
@@ -272,7 +272,7 @@ class Lfm2AudioARForConditionalGeneration(nn.Module):
         assert self._pending_hidden is not None, "compute_logits must run before sample"
         hidden = self._pending_hidden[row]
         frame = self.audio_head.sample_frame(
-            hidden.float(),
+            hidden,
             temperature=self.audio_temperature,
             top_k=self.audio_top_k,
         )
